+++ ParallaxGenLib/include/patchers/base/PatcherMeshShaderTransform.hpp
@@ -4,14 +4,14 @@
 #include <unordered_set>
 
 #include "NIFUtil.hpp"
-#include "patchers/PatcherMesh.hpp"
-#include "patchers/PatcherShader.hpp"
+#include "patchers/base/PatcherMesh.hpp"
+#include "patchers/base/PatcherMeshShader.hpp"
 
 /**
- * @class PatcherShaderTransform
+ * @class PatcherMeshShaderTransform
  * @brief Base class for shader transform patchers
  */
-class PatcherShaderTransform : public PatcherMesh {
+class PatcherMeshShaderTransform : public PatcherMesh {
 private:
     struct ErrorTrackerHasher {
         auto operator()(const std::tuple<std::filesystem::path, NIFUtil::ShapeShader, NIFUtil::ShapeShader>& key) const
@@ -36,18 +36,18 @@ protected:
 
 public:
     // Custom type definitions
-    using PatcherShaderTransformFactory
-        = std::function<std::unique_ptr<PatcherShaderTransform>(std::filesystem::path, nifly::NifFile*)>;
-    using PatcherShaderTransformObject = std::unique_ptr<PatcherShaderTransform>;
+    using PatcherMeshShaderTransformFactory
+        = std::function<std::unique_ptr<PatcherMeshShaderTransform>(std::filesystem::path, nifly::NifFile*)>;
+    using PatcherMeshShaderTransformObject = std::unique_ptr<PatcherMeshShaderTransform>;
 
     // Constructors
-    PatcherShaderTransform(std::filesystem::path nifPath, nifly::NifFile* nif, std::string patcherName,
+    PatcherMeshShaderTransform(std::filesystem::path nifPath, nifly::NifFile* nif, std::string patcherName,
         const NIFUtil::ShapeShader& from, const NIFUtil::ShapeShader& to);
-    virtual ~PatcherShaderTransform() = default;
-    PatcherShaderTransform(const PatcherShaderTransform& other) = default;
-    auto operator=(const PatcherShaderTransform& other) -> PatcherShaderTransform& = default;
-    PatcherShaderTransform(PatcherShaderTransform&& other) noexcept = default;
-    auto operator=(PatcherShaderTransform&& other) noexcept -> PatcherShaderTransform& = default;
+    virtual ~PatcherMeshShaderTransform() = default;
+    PatcherMeshShaderTransform(const PatcherMeshShaderTransform& other) = default;
+    auto operator=(const PatcherMeshShaderTransform& other) -> PatcherMeshShaderTransform& = default;
+    PatcherMeshShaderTransform(PatcherMeshShaderTransform&& other) noexcept = default;
+    auto operator=(PatcherMeshShaderTransform&& other) noexcept -> PatcherMeshShaderTransform& = default;
 
     /**
      * @brief Transform shader match to new shader match
@@ -55,6 +55,7 @@ public:
      * @param FromMatch shader match to transform
      * @return PatcherShader::PatcherMatch transformed match
      */
-    virtual auto transform(const PatcherShader::PatcherMatch& fromMatch, PatcherShader::PatcherMatch& result) -> bool
+    virtual auto transform(const PatcherMeshShader::PatcherMatch& fromMatch, PatcherMeshShader::PatcherMatch& result)
+        -> bool
         = 0;
 };