+++ ParallaxGenLib/src/patchers/base/PatcherMesh.cpp
@@ -1,4 +1,4 @@
-#include "patchers/PatcherMesh.hpp"
+#include "patchers/base/PatcherMesh.hpp"
 
 using namespace std;