+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.8.8] - 2025-02-04
+
+- Added "Fix Mesh Lighting" pre-patcher (Thanks to Catnyss for their article!)
+- Fixed bug where texture sets of duplicate NIFs were not being patched even if an alternate texture didn't exist
+- Changed alpha message to say "beta"
+
 ## [0.8.7] - 2025-02-02
 
 - Added --full-dump CLI argument to ParallaxGen.exe to generate a full crash dump